+++ src/components/Login.vue
@@ -1,0 +1,7 @@
+<script setup>
+import { ref } from "vue";
+</script>
+
+<template>
+    <h1>Login Page</h1>
+</template>
